src/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
anyio==3.2.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
fastapi==0.66.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
praw==7.3.0
prawcore==2.2.0
pydantic==1.8.2
python-dotenv==0.18.0
python-multipart==0.0.5
regex==2021.7.1
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
uvicorn==0.14.0
websocket-client==1.1.0
yarl==1.6.3
